+++ single-header/ctre.hpp
@@ -4440,6 +4440,8 @@ constexpr CTRE_FORCE_INLINE R evaluate(const Iterator begin, Iterator current, c
 #ifndef CTRE_V2__CTRE__ITERATOR__HPP
 #define CTRE_V2__CTRE__ITERATOR__HPP
 
+#include <cstddef>
+
 namespace ctre {
 
 // TODO make proper iterator traits here
@@ -4453,7 +4455,7 @@ template <typename BeginIterator, typename EndIterator, typename RE, typename Re
 	using iterator_category = std::forward_iterator_tag;
 	using pointer = void;
 	using reference = const value_type &;
-	using difference_type = ssize_t;
+	using difference_type = int;
 	
 	BeginIterator orig_begin{};
 	BeginIterator current{};
@@ -4531,7 +4533,7 @@ template <typename BeginIterator, typename EndIterator, typename RE, typename Re
 	using iterator_category = std::forward_iterator_tag;
 	using pointer = void;
 	using reference = const value_type &;
-	using difference_type = ssize_t;
+	using difference_type = int;
 
 	BeginIterator orig_begin{};
 	BeginIterator current{};
@@ -4675,7 +4677,7 @@ template <typename Range, typename RE> struct multi_subject_range {
 		using iterator_category = std::forward_iterator_tag;
 		using pointer = void;
 		using reference = const value_type &;
-		using difference_type = ssize_t;
+		using difference_type = int;
 		
 		first_type first{};
 		last_type last{};
